requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.9
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
urllib3==1.25.8
